# feat(Define) remove overhead on lazy methods (?)