--- a/lib/graphql/argument.rb
+++ b/lib/graphql/argument.rb
@@ -17,24 +17,35 @@ module GraphQL
   class Argument
     include GraphQL::Define::InstanceDefinable
     accepts_definitions :name, :type, :description, :default_value
-    lazy_defined_attr_accessor :type, :description, :default_value, :name
+
+    lazy_methods do
+      attr_accessor :description, :default_value
+
+      def name
+        @name
+      end
+
+      def name=(new_name)
+        @name = new_name
+      end
+
+      # @param new_input_type [GraphQL::BaseType, Proc] Assign a new input type for this argument (if it's a proc, it will be called after schema initialization)
+      def type=(new_input_type)
+        ensure_defined
+        @clean_type = nil
+        @dirty_type = new_input_type
+      end
+
+      # @return [GraphQL::BaseType] the input type for this argument
+      def type
+        @clean_type ||= begin
+          ensure_defined
+          GraphQL::BaseType.resolve_related_type(@dirty_type)
+        end
+      end
+    end
 
     # @!attribute name
     #   @return [String] The name of this argument on its {GraphQL::Field} or {GraphQL::InputObjectType}
-
-    # @param new_input_type [GraphQL::BaseType, Proc] Assign a new input type for this argument (if it's a proc, it will be called after schema initialization)
-    def type=(new_input_type)
-      ensure_defined
-      @clean_type = nil
-      @dirty_type = new_input_type
-    end
-
-    # @return [GraphQL::BaseType] the input type for this argument
-    def type
-      @clean_type ||= begin
-        ensure_defined
-        GraphQL::BaseType.resolve_related_type(@dirty_type)
-      end
-    end
   end
 end