--- a/lib/graphql/define/instance_definable.rb
+++ b/lib/graphql/define/instance_definable.rb
@@ -91,6 +91,7 @@ module GraphQL
         # make sure the previous definition_proc was executed:
         ensure_defined
         @pending_definition = Definition.new(kwargs, block)
+        LazyMethodProxy.create(self, self.class.lazy_definition)
         nil
       end
 
@@ -125,7 +126,6 @@ module GraphQL
       def applied_definitions
         @applied_definitions ||= []
       end
-
 
       class Definition
         def initialize(define_keywords, define_proc)
@@ -174,23 +174,6 @@ module GraphQL
           @own_dictionary = own_dictionary.merge(new_assignments)
         end
 
-        # Define a reader and writer for each of `attr_names` which
-        # ensures that the definition block was called before accessing it.
-        def lazy_defined_attr_accessor(*attr_names)
-          attr_names.each do |attr_name|
-            ivar_name = :"@#{attr_name}"
-            define_method(attr_name) do
-              ensure_defined
-              instance_variable_get(ivar_name)
-            end
-
-            define_method("#{attr_name}=") do |new_value|
-              ensure_defined
-              instance_variable_set(ivar_name, new_value)
-            end
-          end
-        end
-
         # @return [Hash] combined definitions for self and ancestors
         def dictionary
           if superclass.respond_to?(:dictionary)
@@ -204,6 +187,19 @@ module GraphQL
         def own_dictionary
           @own_dictionary ||= {}
         end
+
+        # The methods defined in this block will ensure that the `.define`
+        # block has been executed before they run.
+        def lazy_methods(&lazy_definition)
+          if @lazy_definition
+            raise("Can't define lazy_methods twice")
+          else
+            @lazy_definition = lazy_definition
+          end
+          LazyMethodProxy.create(self, @lazy_definition)
+        end
+
+        attr_reader :lazy_definition
       end
 
       class AssignMetadataKey