--- a/lib/graphql/field.rb
+++ b/lib/graphql/field.rb
@@ -125,12 +125,8 @@ module GraphQL
       :property, :hash_key, :complexity, :mutation,
       argument: GraphQL::Define::AssignArgument
 
-
-    lazy_defined_attr_accessor :name, :deprecation_reason, :description, :property, :hash_key, :mutation, :arguments, :complexity
-
     # @!attribute [r] resolve_proc
     #   @return [<#call(obj, args,ctx)>] A proc-like object which can be called to return the field's value
-    attr_reader :resolve_proc
 
     # @!attribute name
     #   @return [String] The name of this field on its {GraphQL::ObjectType} (or {GraphQL::InterfaceType})
@@ -147,72 +143,67 @@ module GraphQL
     def initialize
       @complexity = 1
       @arguments = {}
-      @resolve_proc = build_default_resolver
     end
 
-    # Get a value for this field
-    # @example resolving a field value
-    #   field.resolve(obj, args, ctx)
-    #
-    # @param object [Object] The object this field belongs to
-    # @param arguments [Hash] Arguments declared in the query
-    # @param context [GraphQL::Query::Context]
-    def resolve(object, arguments, context)
-      ensure_defined
-      resolve_proc.call(object, arguments, context)
-    end
+    lazy_methods do
+      attr_accessor :deprecation_reason, :description, :property, :hash_key, :mutation, :arguments, :complexity
+      attr_reader :name
 
-    # Provide a new callable for this field's resolve function. If `nil`,
-    # a new resolve proc will be build based on its {#name}, {#property} or {#hash_key}.
-    # @param new_resolve_proc [<#call(obj, args, ctx)>, nil]
-    def resolve=(new_resolve_proc)
-      ensure_defined
-      @resolve_proc = new_resolve_proc || build_default_resolver
-    end
-
-    def type=(new_return_type)
-      ensure_defined
-      @clean_type = nil
-      @dirty_type = new_return_type
-    end
-
-    # Get the return type for this field.
-    def type
-      @clean_type ||= begin
-        ensure_defined
-        GraphQL::BaseType.resolve_related_type(@dirty_type)
+      def resolve_proc
+        @resolve_proc ||= build_default_resolver
       end
-    end
 
-    # You can only set a field's name _once_ -- this to prevent
-    # passing the same {Field} to multiple `.field` calls.
-    #
-    # This is important because {#name} may be used by {#resolve}.
-    def name=(new_name)
-      ensure_defined
-      if @name.nil?
-        @name = new_name
-      elsif @name != new_name
-        raise("Can't rename an already-named field. (Tried to rename \"#{@name}\" to \"#{new_name}\".) If you're passing a field with the `field:` argument, make sure it's an unused instance of GraphQL::Field.")
+      # Get a value for this field
+      # @example resolving a field value
+      #   field.resolve(obj, args, ctx)
+      #
+      # @param object [Object] The object this field belongs to
+      # @param arguments [Hash] Arguments declared in the query
+      # @param context [GraphQL::Query::Context]
+      def resolve(object, arguments, context)
+        resolve_proc.call(object, arguments, context)
       end
-    end
 
-    # @param new_property [Symbol] A method to call to resolve this field. Overrides the existing resolve proc.
-    def property=(new_property)
-      ensure_defined
-      @property = new_property
-      self.resolve = nil # reset resolve proc
-    end
+      # Provide a new callable for this field's resolve function. If `nil`,
+      # a new resolve proc will be build based on its {#name}, {#property} or {#hash_key}.
+      # @param new_resolve_proc [<#call(obj, args, ctx)>, nil]
+      def resolve=(new_resolve_proc)
+        @resolve_proc = new_resolve_proc || build_default_resolver
+      end
 
-    # @param new_hash_key [Symbol] A key to access with `#[key]` to resolve this field. Overrides the existing resolve proc.
-    def hash_key=(new_hash_key)
-      ensure_defined
-      @hash_key = new_hash_key
-      self.resolve = nil # reset resolve proc
-    end
+      def type=(new_return_type)
+        @clean_type = nil
+        @dirty_type = new_return_type
+      end
 
-    def to_s
-      "<Field name:#{name || "not-named"} desc:#{description} resolve:#{resolve_proc}>"
+      # Get the return type for this field.
+      def type
+        @clean_type ||= GraphQL::BaseType.resolve_related_type(@dirty_type)
+      end
+
+      # You can only set a field's name _once_ -- this to prevent
+      # passing the same {Field} to multiple `.field` calls.
+      #
+      # This is important because {#name} may be used by {#resolve}.
+      def name=(new_name)
+        if @name.nil?
+          @name = new_name
+        elsif @name != new_name
+          raise("Can't rename an already-named field. (Tried to rename \"#{@name}\" to \"#{new_name}\".) If you're passing a field with the `field:` argument, make sure it's an unused instance of GraphQL::Field.")
+        end
+      end
+
+      # @param new_property [Symbol] A method to call to resolve this field. Overrides the existing resolve proc.
+      def property=(new_property)
+        @property = new_property
+        self.resolve = nil # reset resolve proc
+      end
+
+      # @param new_hash_key [Symbol] A key to access with `#[key]` to resolve this field. Overrides the existing resolve proc.
+      def hash_key=(new_hash_key)
+        @hash_key = new_hash_key
+        self.resolve = nil # reset resolve proc
+      end
     end
 
     private