--- a/lib/graphql/union_type.rb
+++ b/lib/graphql/union_type.rb
@@ -38,14 +38,14 @@ module GraphQL
       @dirty_possible_types = new_possible_types
     end
 
-    def possible_types
-      @clean_possible_types ||= begin
-        ensure_defined
-
-        if @dirty_possible_types.respond_to?(:map)
-          @dirty_possible_types.map { |type| GraphQL::BaseType.resolve_related_type(type) }
-        else
-          @dirty_possible_types
+    lazy_methods do
+      def possible_types
+        @clean_possible_types ||= begin
+          if @dirty_possible_types.respond_to?(:map)
+            @dirty_possible_types.map { |type| GraphQL::BaseType.resolve_related_type(type) }
+          else
+            @dirty_possible_types
+          end
         end
       end
     end