I thought I'd take a look at a query run to see where the time goes. Here's my profiling script:

<details>

``` ruby
# introspection_profile.rb 
require "graphql"
require "dairy_app"
require "ruby-prof"

prof_name = "./prof-#{`git rev-parse --abbrev-ref HEAD`}"
document = GraphQL.parse(GraphQL::Introspection::INTROSPECTION_QUERY)

def run_query(doc)
  DummySchema.execute(document: doc)
end

# Make sure it's warm
run_query(document)

RubyProf.start

run_query(document)

result = RubyProf.stop
File.open(prof_name, "wb") { |f| RubyProf::FlatPrinter.new(result).print(f) }
```

</details>

On master, I got: 

<details>

```
~/code/graphql $ head -n 15 prof-master
Measure Mode: wall_time
Thread ID: 70287614024120
Fiber ID: 70287614744860
Total: 0.155563
Sort by: self_time

 %self      total      self      wait     child     calls  name
  8.36      0.040     0.013     0.000     0.027     8662  *Class#new
  7.37      0.012     0.011     0.000     0.000     9540  *GraphQL::Define::InstanceDefinable#ensure_defined
  5.42      0.008     0.008     0.000     0.000     6821   Kernel#instance_variable_get
  5.02      0.008     0.008     0.000     0.000     8625   Kernel#hash
  3.49      0.027     0.005     0.000     0.021     3358  *GraphQL::Schema::MiddlewareChain#call
  3.36      0.052     0.005     0.000     0.047     1679   GraphQL::Query::SerialExecution::FieldResolution#get_raw_value
  3.18      0.013     0.005     0.000     0.008     3526   GraphQL::BaseType#name
  2.87      0.004     0.004     0.000     0.000     2810   GraphQL::Query::SerialExecution::ValueResolution::BaseResolution#initialize
```

</details>

This one stood out to me: 

```
  7.37      0.012     0.011     0.000     0.000     9540  *GraphQL::Define::InstanceDefinable#ensure_defined
```

Yikes, we lose _7%_ of our time to overhead on lazy definitions? 

I wonder if I can remove the overhead. I heard of a technique for applying a debugger in JRuby+Truffle where you swap entire method bodies when the debugger is activated. When it's deactivated, you swap back, so you're left with the original method definition with no overhead. 

Could we do something similar? On initialization, all the object's methods are "empty". But when you first call one of them, the methods call the lazy definition and replace themselves (in this case, override themselves) with the real implementations.

Maybe. Here's the profile result on this branch: 

<details>

```
~/code/graphql $ git co lazy-method-proxy
~/code/graphql $ ruby -Ilib:spec/support introspection_profile.rb
~/code/graphql $ head -n 15 prof-lazy-method-proxy
Measure Mode: wall_time
Thread ID: 70104960473560
Fiber ID: 70104962774120
Total: 0.138191
Sort by: self_time

 %self      total      self      wait     child     calls  name
 11.05      0.045     0.015     0.000     0.030     8661  *Class#new
  7.16      0.010     0.010     0.000     0.000     8625   Kernel#hash
  4.07      0.044     0.006     0.000     0.038     1679   GraphQL::Query::SerialExecution::FieldResolution#get_raw_value
  3.53      0.021     0.005     0.000     0.016     3358  *GraphQL::Schema::MiddlewareChain#call
  3.27      0.005     0.005     0.000     0.000     2810   GraphQL::Query::SerialExecution::ValueResolution::BaseResolution#initialize
  3.03      0.004     0.004     0.000     0.000     3360   GraphQL::BaseType#to_s
  2.90      0.004     0.004     0.000     0.000     3561   #<Module:0x007f852b12ac70>#type
  2.51      0.014     0.003     0.000     0.011     1679   GraphQL::Query::SerialExecution::FieldResolution#initialize
```

</details>

Sad thing is, the benchmark isn't any better: 

<details>

```
~/code/graphql $ cat introspection_benchmark.rb
require "graphql"
require "dairy_app"
require "benchmark/ips"

document = GraphQL.parse(GraphQL::Introspection::INTROSPECTION_QUERY)

def run_query(doc)
  DummySchema.execute(document: doc)
end

run_query(document)

Benchmark.ips do |x|
  x.report("run query") { run_query(document) }
  x.compare!
end
~/code/graphql $ git co master
Switched to branch 'master'
Your branch is up-to-date with 'origin/master'.
~/code/graphql $ ruby -Ilib:spec/support introspection_benchmark.rb
Warming up --------------------------------------
           run query     3.000  i/100ms
Calculating -------------------------------------
           run query     33.035  (± 6.1%) i/s -    165.000  in   5.012511s
~/code/graphql $ ruby -Ilib:spec/support introspection_benchmark.rb
Warming up --------------------------------------
           run query     3.000  i/100ms
Calculating -------------------------------------
           run query     33.569  (± 3.0%) i/s -    168.000  in   5.010186s
~/code/graphql $ ruby -Ilib:spec/support introspection_benchmark.rb
Warming up --------------------------------------
           run query     3.000  i/100ms
Calculating -------------------------------------
           run query     34.527  (± 2.9%) i/s -    174.000  in   5.043941s
~/code/graphql $ git co lazy-method-proxy
Switched to branch 'lazy-method-proxy'
~/code/graphql $ ruby -Ilib:spec/support introspection_benchmark.rb
Warming up --------------------------------------
           run query     3.000  i/100ms
Calculating -------------------------------------
           run query     35.062  (± 2.9%) i/s -    177.000  in   5.052376s
~/code/graphql $ ruby -Ilib:spec/support introspection_benchmark.rb
Warming up --------------------------------------
           run query     3.000  i/100ms
Calculating -------------------------------------
           run query     33.842  (± 3.0%) i/s -    171.000  in   5.055781s
~/code/graphql $ ruby -Ilib:spec/support introspection_benchmark.rb
Warming up --------------------------------------
           run query     3.000  i/100ms
Calculating -------------------------------------
           run query     34.488  (± 5.8%) i/s -    174.000  in   5.060149s
```

</details>

So, unless I'm missing something, maybe it's not worth all the hand-waving after all!