@rmosolgo 2016-10-31
I'm optimistic about this idea but I don't like that we lose all the yardoc. I'm going to try again another way some other time.